Bug 1569828 - Make confirm-delete-dialog a reusable component. r=fluent-reviewers,jaws,flod
Differential Revision: https://phabricator.services.mozilla.com/D39819

UltraBlame original commit: fe3a63d0b9755e2385c2e29e67a5fea18c60b05a

+++ browser/components/aboutlogins/content/components/login-item.js
@@ -263,20 +263,38 @@ export default class LoginItem extends HTMLElement {
 
 
 
+
+  showConfirmationDialog(type, onConfirm = () => {}) {
+    const dialog = document.querySelector("confirmation-dialog");
+    let options;
+    switch (type) {
+      case "delete": {
+        options = {
+          title: "confirm-delete-dialog-title",
+          message: "confirm-delete-dialog-message",
+          confirmButtonLabel: "confirm-delete-dialog-confirm-button",
+        };
+      }
+    }
+    let dialogPromise = dialog.show(options);
+    dialogPromise.then(onConfirm, () => {});
+    return dialogPromise;
+  }
+
+  
+
+
+
   confirmDelete() {
-    const dialog = document.querySelector("confirm-delete-dialog");
-    dialog.show().then(
-      () => {
-        document.dispatchEvent(
-          new CustomEvent("AboutLoginsDeleteLogin", {
-            bubbles: true,
-            detail: this._login,
-          })
-        );
-        recordTelemetryEvent({ object: "existing_login", method: "delete" });
-      },
-      () => {}
-    );
+    this.showConfirmationDialog("delete", () => {
+      document.dispatchEvent(
+        new CustomEvent("AboutLoginsDeleteLogin", {
+          bubbles: true,
+          detail: this._login,
+        })
+      );
+      recordTelemetryEvent({ object: "existing_login", method: "delete" });
+    });
   }
